[Bugfix] Allow vllm to still work if triton is not installed. (#6786)
